+++ ot/gromov/_gw.py
@@ -16,14 +16,14 @@ import numpy as np
 
 from ..utils import dist, UndefinedParameter, list_to_array
 from ..optim import cg, line_search_armijo, solve_1d_linesearch_quad
-from ..utils import check_random_state
+from ..utils import check_random_state, unif
 from ..backend import get_backend, NumpyBackend
 
 from ._utils import init_matrix, gwloss, gwggrad
 from ._utils import update_square_loss, update_kl_loss, update_feature_matrix
 
 
-def gromov_wasserstein(C1, C2, p, q, loss_fun='square_loss', symmetric=None, log=False, armijo=False, G0=None,
+def gromov_wasserstein(C1, C2, p=None, q=None, loss_fun='square_loss', symmetric=None, log=False, armijo=False, G0=None,
                        max_iter=1e4, tol_rel=1e-9, tol_abs=1e-9, **kwargs):
     r"""
     Returns the Gromov-Wasserstein transport between :math:`(\mathbf{C_1}, \mathbf{p})` and :math:`(\mathbf{C_2}, \mathbf{q})`
@@ -60,11 +60,13 @@ def gromov_wasserstein(C1, C2, p, q, loss_fun='square_loss', symmetric=None, log
         Metric cost matrix in the source space
     C2 : array-like, shape (nt, nt)
         Metric cost matrix in the target space
-    p : array-like, shape (ns,)
-        Distribution in the source space
-    q : array-like, shape (nt,)
-        Distribution in the target space
-    loss_fun : str
+    p : array-like, shape (ns,), optional
+        Distribution in the source space.
+        If let to its default value None, uniform distribution is taken.
+    q : array-like, shape (nt,), optional
+        Distribution in the target space.
+        If let to its default value None, uniform distribution is taken.
+    loss_fun : str, optional
         loss function used for the solver either 'square_loss' or 'kl_loss'
     symmetric : bool, optional
         Either C1 and C2 are to be assumed symmetric or not.
@@ -112,15 +114,24 @@ def gromov_wasserstein(C1, C2, p, q, loss_fun='square_loss', symmetric=None, log
         distance between networks and stable network invariants.
         Information and Inference: A Journal of the IMA, 8(4), 757-787.
     """
-    p, q = list_to_array(p, q)
-    p0, q0, C10, C20 = p, q, C1, C2
-    if G0 is None:
-        nx = get_backend(p0, q0, C10, C20)
+    arr = [C1, C2]
+    if p is not None:
+        arr.append(list_to_array(p))
     else:
+        p = unif(C1.shape[0], type_as=C1)
+    if q is not None:
+        arr.append(list_to_array(q))
+    else:
+        q = unif(C2.shape[0], type_as=C2)
+    if G0 is not None:
         G0_ = G0
-        nx = get_backend(p0, q0, C10, C20, G0_)
-    p = nx.to_numpy(p)
-    q = nx.to_numpy(q)
+        arr.append(G0)
+
+    nx = get_backend(*arr)
+    p0, q0, C10, C20 = p, q, C1, C2
+
+    p = nx.to_numpy(p0)
+    q = nx.to_numpy(q0)
     C1 = nx.to_numpy(C10)
     C2 = nx.to_numpy(C20)
     if symmetric is None:
@@ -168,7 +179,7 @@ def gromov_wasserstein(C1, C2, p, q, loss_fun='square_loss', symmetric=None, log
         return nx.from_numpy(cg(p, q, 0., 1., f, df, G0, line_search, log=False, numItermax=max_iter, stopThr=tol_rel, stopThr2=tol_abs, **kwargs), type_as=C10)
 
 
-def gromov_wasserstein2(C1, C2, p, q, loss_fun='square_loss', symmetric=None, log=False, armijo=False, G0=None,
+def gromov_wasserstein2(C1, C2, p=None, q=None, loss_fun='square_loss', symmetric=None, log=False, armijo=False, G0=None,
                         max_iter=1e4, tol_rel=1e-9, tol_abs=1e-9, **kwargs):
     r"""
     Returns the Gromov-Wasserstein discrepancy between :math:`(\mathbf{C_1}, \mathbf{p})` and :math:`(\mathbf{C_2}, \mathbf{q})`
@@ -209,10 +220,12 @@ def gromov_wasserstein2(C1, C2, p, q, loss_fun='square_loss', symmetric=None, lo
         Metric cost matrix in the source space
     C2 : array-like, shape (nt, nt)
         Metric cost matrix in the target space
-    p : array-like, shape (ns,)
+    p : array-like, shape (ns,), optional
         Distribution in the source space.
-    q :  array-like, shape (nt,)
+        If let to its default value None, uniform distribution is taken.
+    q : array-like, shape (nt,), optional
         Distribution in the target space.
+        If let to its default value None, uniform distribution is taken.
     loss_fun :  str
         loss function used for the solver either 'square_loss' or 'kl_loss'
     symmetric : bool, optional
@@ -266,6 +279,12 @@ def gromov_wasserstein2(C1, C2, p, q, loss_fun='square_loss', symmetric=None, lo
     # simple get_backend as the full one will be handled in gromov_wasserstein
     nx = get_backend(C1, C2)
 
+    # init marginals if set as None
+    if p is None:
+        p = unif(C1.shape[0], type_as=C1)
+    if q is None:
+        q = unif(C2.shape[0], type_as=C2)
+
     T, log_gw = gromov_wasserstein(
         C1, C2, p, q, loss_fun, symmetric, log=True, armijo=armijo, G0=G0,
         max_iter=max_iter, tol_rel=tol_rel, tol_abs=tol_abs, **kwargs)
@@ -286,7 +305,7 @@ def gromov_wasserstein2(C1, C2, p, q, loss_fun='square_loss', symmetric=None, lo
         return gw
 
 
-def fused_gromov_wasserstein(M, C1, C2, p, q, loss_fun='square_loss', symmetric=None, alpha=0.5,
+def fused_gromov_wasserstein(M, C1, C2, p=None, q=None, loss_fun='square_loss', symmetric=None, alpha=0.5,
                              armijo=False, G0=None, log=False, max_iter=1e4, tol_rel=1e-9, tol_abs=1e-9, **kwargs):
     r"""
     Computes the FGW transport between two graphs (see :ref:`[24] <references-fused-gromov-wasserstein>`)
@@ -323,10 +342,12 @@ def fused_gromov_wasserstein(M, C1, C2, p, q, loss_fun='square_loss', symmetric=
         Metric cost matrix representative of the structure in the source space
     C2 : array-like, shape (nt, nt)
         Metric cost matrix representative of the structure in the target space
-    p : array-like, shape (ns,)
-        Distribution in the source space
-    q : array-like, shape (nt,)
-        Distribution in the target space
+    p : array-like, shape (ns,), optional
+        Distribution in the source space.
+        If let to its default value None, uniform distribution is taken.
+    q : array-like, shape (nt,), optional
+        Distribution in the target space.
+        If let to its default value None, uniform distribution is taken.
     loss_fun : str, optional
         Loss function used for the solver
     symmetric : bool, optional
@@ -372,16 +393,24 @@ def fused_gromov_wasserstein(M, C1, C2, p, q, loss_fun='square_loss', symmetric=
         distance between networks and stable network invariants.
         Information and Inference: A Journal of the IMA, 8(4), 757-787.
     """
-    p, q = list_to_array(p, q)
-    p0, q0, C10, C20, M0, alpha0 = p, q, C1, C2, M, alpha
-    if G0 is None:
-        nx = get_backend(p0, q0, C10, C20, M0)
+    arr = [C1, C2, M]
+    if p is not None:
+        arr.append(list_to_array(p))
     else:
+        p = unif(C1.shape[0], type_as=C1)
+    if q is not None:
+        arr.append(list_to_array(q))
+    else:
+        q = unif(C2.shape[0], type_as=C2)
+    if G0 is not None:
         G0_ = G0
-        nx = get_backend(p0, q0, C10, C20, M0, G0_)
+        arr.append(G0)
 
-    p = nx.to_numpy(p)
-    q = nx.to_numpy(q)
+    nx = get_backend(*arr)
+    p0, q0, C10, C20, M0, alpha0 = p, q, C1, C2, M, alpha
+
+    p = nx.to_numpy(p0)
+    q = nx.to_numpy(q0)
     C1 = nx.to_numpy(C10)
     C2 = nx.to_numpy(C20)
     M = nx.to_numpy(M0)
@@ -433,7 +462,7 @@ def fused_gromov_wasserstein(M, C1, C2, p, q, loss_fun='square_loss', symmetric=
         return nx.from_numpy(cg(p, q, (1 - alpha) * M, alpha, f, df, G0, line_search, log=False, numItermax=max_iter, stopThr=tol_rel, stopThr2=tol_abs, **kwargs), type_as=C10)
 
 
-def fused_gromov_wasserstein2(M, C1, C2, p, q, loss_fun='square_loss', symmetric=None, alpha=0.5,
+def fused_gromov_wasserstein2(M, C1, C2, p=None, q=None, loss_fun='square_loss', symmetric=None, alpha=0.5,
                               armijo=False, G0=None, log=False, max_iter=1e4, tol_rel=1e-9, tol_abs=1e-9, **kwargs):
     r"""
     Computes the FGW distance between two graphs see (see :ref:`[24] <references-fused-gromov-wasserstein2>`)
@@ -474,10 +503,12 @@ def fused_gromov_wasserstein2(M, C1, C2, p, q, loss_fun='square_loss', symmetric
         Metric cost matrix representative of the structure in the source space.
     C2 : array-like, shape (nt, nt)
         Metric cost matrix representative of the structure in the target space.
-    p :  array-like, shape (ns,)
+    p : array-like, shape (ns,), optional
         Distribution in the source space.
-    q :  array-like, shape (nt,)
+        If let to its default value None, uniform distribution is taken.
+    q : array-like, shape (nt,), optional
         Distribution in the target space.
+        If let to its default value None, uniform distribution is taken.
     loss_fun : str, optional
         Loss function used for the solver.
     symmetric : bool, optional
@@ -528,6 +559,12 @@ def fused_gromov_wasserstein2(M, C1, C2, p, q, loss_fun='square_loss', symmetric
         Information and Inference: A Journal of the IMA, 8(4), 757-787.
     """
     nx = get_backend(C1, C2, M)
+
+    # init marginals if set as None
+    if p is None:
+        p = unif(C1.shape[0], type_as=C1)
+    if q is None:
+        q = unif(C2.shape[0], type_as=C2)
 
     T, log_fgw = fused_gromov_wasserstein(
         M, C1, C2, p, q, loss_fun, symmetric, alpha, armijo, G0, log=True,
@@ -627,7 +664,7 @@ def solve_gromov_linesearch(G, deltaG, cost_G, C1, C2, M, reg,
 
 
 def gromov_barycenters(
-        N, Cs, ps, p, lambdas, loss_fun, symmetric=True, armijo=False,
+        N, Cs, ps=None, p=None, lambdas=None, loss_fun='square_loss', symmetric=True, armijo=False,
         max_iter=1000, tol=1e-9, warmstartT=False, verbose=False, log=False,
         init_C=None, random_state=None, **kwargs):
     r"""
@@ -650,13 +687,16 @@ def gromov_barycenters(
         Size of the targeted barycenter
     Cs : list of S array-like of shape (ns, ns)
         Metric cost matrices
-    ps : list of S array-like of shape (ns,)
-        Sample weights in the `S` spaces
-    p : array-like, shape (N,)
-        Weights in the targeted barycenter
-    lambdas : list of float
-        List of the `S` spaces' weights
-    loss_fun : callable
+    ps : list of S array-like of shape (ns,), optional
+        Sample weights in the `S` spaces.
+        If let to its default value None, uniform distributions are taken.
+    p : array-like, shape (N,), optional
+        Weights in the targeted barycenter.
+        If let to its default value None, uniform distribution is taken.
+    lambdas : list of float, optional
+        List of the `S` spaces' weights.
+        If let to its default value None, uniform weights are taken.
+    loss_fun : callable, optional
         tensor-matrix multiplication function based on specific loss function
     symmetric : bool, optional.
         Either structures are to be assumed symmetric or not. Default value is True.
@@ -696,11 +736,21 @@ def gromov_barycenters(
 
     """
     Cs = list_to_array(*Cs)
-    ps = list_to_array(*ps)
-    p = list_to_array(p)
-    nx = get_backend(*Cs, *ps, p)
+    arr = [*Cs]
+    if ps is not None:
+        arr += list_to_array(*ps)
+    else:
+        ps = [unif(C.shape[0], type_as=C) for C in Cs]
+    if p is not None:
+        arr.append(list_to_array(p))
+    else:
+        p = unif(N, type_as=Cs[0])
+
+    nx = get_backend(*arr)
 
     S = len(Cs)
+    if lambdas is None:
+        lambdas = [1. / S] * S
 
     # Initialization of C : random SPD matrix (if not provided by user)
     if init_C is None:
@@ -758,7 +808,7 @@ def gromov_barycenters(
 
 
 def fgw_barycenters(
-        N, Ys, Cs, ps, lambdas, alpha, fixed_structure=False,
+        N, Ys, Cs, ps=None, lambdas=None, alpha=0.5, fixed_structure=False,
         fixed_features=False, p=None, loss_fun='square_loss', armijo=False,
         symmetric=True, max_iter=100, tol=1e-9, warmstartT=False, verbose=False,
         log=False, init_C=None, init_X=None, random_state=None, **kwargs):
@@ -772,16 +822,21 @@ def fgw_barycenters(
         Features of all samples
     Cs : list of array-like, each element has shape (ns,ns)
         Structure matrices of all samples
-    ps : list of array-like, each element has shape (ns,)
+    ps : list of array-like, each element has shape (ns,), optional
         Masses of all samples.
-    lambdas : list of float
-        List of the `S` spaces' weights
-    alpha : float
-        Alpha parameter for the fgw distance
+        If let to its default value None, uniform distributions are taken.
+    lambdas : list of float, optional
+        List of the `S` spaces' weights.
+        If let to its default value None, uniform weights are taken.
+    alpha : float, optional
+        Alpha parameter for the fgw distance.
     fixed_structure : bool
         Whether to fix the structure of the barycenter during the updates
     fixed_features : bool
         Whether to fix the feature of the barycenter during the updates
+    p : array-like, shape (N,), optional
+        Weights in the targeted barycenter.
+        If let to its default value None, uniform distribution is taken.
     loss_fun : str
         Loss function used for the solver either 'square_loss' or 'kl_loss'
     symmetric : bool, optional
@@ -829,15 +884,24 @@ def fgw_barycenters(
         International Conference on Machine Learning (ICML). 2019.
     """
     Cs = list_to_array(*Cs)
-    ps = list_to_array(*ps)
     Ys = list_to_array(*Ys)
-    p = list_to_array(p)
-    nx = get_backend(*Cs, *Ys, *ps)
+    arr = [*Cs, *Ys]
+    if ps is not None:
+        arr += list_to_array(*ps)
+    else:
+        ps = [unif(C.shape[0], type_as=C) for C in Cs]
+    if p is not None:
+        arr.append(list_to_array(p))
+    else:
+        p = unif(N, type_as=Cs[0])
+
+    nx = get_backend(*arr)
 
     S = len(Cs)
+    if lambdas is None:
+        lambdas = [1. / S] * S
+
     d = Ys[0].shape[1]  # dimension on the node features
-    if p is None:
-        p = nx.ones(N, type_as=Cs[0]) / N
 
     if fixed_structure:
         if init_C is None: